TestAndValidationAgent/src/TestAndValidationAgent/agent.py

from autogen import AssistantAgent
from agoric_sdk import ContractAPI

class TestAndValidationAgent:
    def __init__(self):
        self.assistant = AssistantAgent("validation_assistant")
        self.contract_api = ContractAPI("https://api.agoric.net")

    def test_smart_contract(self, contract_code):
        try:
            # Test the smart contract code
            result = self.contract_api.test_contract(contract_code)
            if result["status"] == "success":
                return "Smart contract test passed."
            else:
                return f"Smart contract test failed: {result['error']}"
        except Exception as e:
            return f"Failed to test smart contract: {str(e)}"

    def validate_application(self, app_code):
        try:
            # Validate the application code
            validation_result = self.assistant.handle_input(f"Validate the following application code: {app_code}")
            return validation_result
        except Exception as e:
            return f"Failed to validate application: {str(e)}"

    def fix_errors(self, code_with_errors):
        try:
            # Fix errors in the provided code
            fixed_code = self.assistant.handle_input(f"Fix the errors in the following code: {code_with_errors}")
            return fixed_code
        except Exception as e:
            return f"Failed to fix errors: {str(e)}"

if __name__ == "__main__":
    agent = TestAndValidationAgent()
    contract_code = "example_smart_contract_code"
    test_result = agent.test_smart_contract(contract_code)
    print(test_result)
    app_code = "example_application_code"
    validation_result = agent.validate_application(app_code)
    print(validation_result)
    code_with_errors = "example_code_with_errors"
    fixed_code = agent.fix_errors(code_with_errors)
    print(fixed_code)
